LayoutTests/canvas/philip/tests/2d.drawImage.null-expected.txt

< [index] >
2d.drawImage.null
References: 2d.drawImage.type
Actual output:
Failed assertion: expected exception of type TYPE_MISMATCH_ERR, got: Type error
